// apps/console/src/api/generated/models/project.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Taskome API
 * Taskome control-plane API
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  /** @nullable */
  archivedAt: string | null;
  createdAt: string;
  /** @nullable */
  description: string | null;
  id: string;
  isDefault: boolean;
  name: string;
  status: ProjectStatus;
  updatedAt: string;
}